Retourne un `ResultatEligibilite` depuis `EnsembleDeSpecifications`
On pousse encore d'un cran l'encapsulation.
`EnsembleDeSpecification` est maintenant en charge de trouver la spécification correspondant aux réponses et d'évaluer son résultat.
Ça va permettre de gérer plus facilement le cas où plusieurs spécifications correspondent.
